refactor: centralize visualization and parsing utilities by moving functions to shared_utils
